llm-ollama-pdf/requirements.txt
ollama_ocr
pdf2image
Pillow 